requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
opencv-python>=4.5.0
streamlit>=1.20.0
ffmpeg-python>=0.2.0
tqdm>=4.65.0
python-dotenv>=1.0.0
basicsr>=1.4.2
facexlib>=0.3.0
realesrgan>=0.3.0
gfpgan>=1.3.8
moviepy>=1.0.3
Pillow>=9.0.0
psutil>=5.9.0
